backend/requirements.txt
Flask>=2.0
python-dotenv
pinecone>=2.2.0
requests
# Optional providers (install at least one)
google-generativeai>=0.3.0    # for Gemini (if you want Gemini)
openai>=1.0.0                 # for OpenAI (optional)
